requirements.txt
mysql-connector-python~=8.0.31
fastapi~=0.88.0
uvicorn~=0.15.0
pydantic~=1.10.2
PyJWT
python-decouple
pydantic[email]